add chat_messages and chat_room routes to chat_routes
