--- server/app/forms.py
@@ -1,30 +1,0 @@
-#!/usr/bin/env python3
-""" Handles registration and login forms """
-
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, BooleanField
-from wtforms.validators import DataRequired, Length, Email, EqualTo
-
-
-class RegistrationForm(FlaskForm):
-    """ Registration form """
-    username = StringField('Username',
-                           validators=[DataRequired(), Length(min=4, max=25)])
-    email = StringField('Email',
-                        validators=[DataRequired(), Email()])
-    password = PasswordField('Password',
-                             validators=[DataRequired(), Length(min=6)])
-    confirm_password = PasswordField('Confirm Password',
-                                     validators=[DataRequired(),
-                                                 EqualTo('password')])
-    submit = SubmitField("Register")
-
-
-class LoginForm(FlaskForm):
-    """ Login form """
-    email = StringField('Email',
-                        validators=[DataRequired(), Email()])
-    password = PasswordField('Password',
-                             validators=[DataRequired(), Length(min=6)])
-    remember_me = BooleanField('Remember Me')
-    submit = SubmitField("Log In")
