--- server/app/templates/index.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <h1>Hi, {{ user.username }}!</h1>
-    <h2>Here are your appointments:</h2>
-{% endblock %}
