--- server/app/templates/login.html
@@ -1,29 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <h3>Login</h3>
-    <form action="" method="POST" novalidate>
-        {{ form.csrf_token }}
-        {{ form.hidden_tag() }}
-        <p>
-            {{ form.email.label }}<br>
-            {{ form.email }}
-            {% for error in form.email.errors %}
-                <span style="color: red;">[{{ error }}]</span>
-            {% endfor %}
-        </p>
-        <p>
-            {{ form.password.label }}<br>
-            {{ form.password }}
-            {% for error in form.password.errors %}
-                <span style="color: red;">[{{ error }}]</span>
-            {% endfor %}
-        </p>
-        <p>
-            {{ form.remember_me() }} {{ form.remember_me.label }}
-        </p>
-        <p>
-            {{ form.submit }}
-        </p>
-    </form>
-{% endblock %}
